refactor and change DI to single tone static
