+++ src/main/java/enums/HttpMethod.java
@@ -1,4 +1,4 @@
-package webserver;
+package enums;
 
 public enum HttpMethod {
     GET,
