+++ src/main/java/enums/StatusCode.java
@@ -1,4 +1,4 @@
-package webserver;
+package enums;
 
 public enum StatusCode {
     OK(200, "OK"),
