+++ src/main/java/enums/URLPath.java
@@ -1,4 +1,4 @@
-package webserver;
+package enums;
 
 
 public enum URLPath {
