+++ src/main/java/webserver/HttpResponse.java
@@ -1,5 +1,7 @@
 package webserver;
 
+import enums.StatusCode;
+
 import java.io.DataOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
