+++ src/main/java/webserver/controller/ForwardController.java
@@ -2,7 +2,6 @@ package webserver.controller;
 
 import webserver.HttpRequest;
 import webserver.HttpResponse;
-import webserver.URLPath;
 
 import java.io.IOException;
